requirements.txt
pandas
fastapi
uvicorn
sklearn.metrics.pairwise